% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkdb.R
\name{close_connection}
\alias{close_connection}
\title{Close connection to kdb+ instance.}
\usage{
close_connection(con)
}
\arguments{
\item{con}{Connection handle.}
}
\value{
0 on closed connection.
}
\description{
Close connection to kdb+ instance.
}
\examples{
\dontrun{
close_connection(h)
}
}
