% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qserver.R
\name{open_connection}
\alias{open_connection}
\title{Connect to kdb+ instance.}
\usage{
open_connection(host = "localhost", port = 5000, user = NULL)
}
\arguments{
\item{host}{Hostname.}

\item{port}{Port number.}

\item{user}{Username and password as user:password string}
}
\value{
Handle to kdb+ instance for \code{execute} and \code{close_connection}.
}
\description{
Connect to kdb+ instance.
}
\examples{
\dontrun{
h<-open_connection()
h<-open_connection(port=5000)
}
}
