% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkdb.R
\name{sslinfo}
\alias{sslinfo}
\title{Show information about openssl configuration found on the system to use for secure connections.}
\usage{
sslinfo()
}
\value{
dictionary with information about ssl.
}
\description{
Show information about openssl configuration found on the system to use for secure connections.
}
\examples{
\dontrun{
sslinfo()
}
}
