% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.qserver.R
\name{test.qserver}
\alias{test.qserver}
\title{Run simple series of tests for qserver
Note: Start kdb+ instance on localhost:5000 before running this function}
\usage{
test.qserver()
}
\description{
Run simple series of tests for qserver
Note: Start kdb+ instance on localhost:5000 before running this function
}
