% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.rkdb.R
\name{test.rkdb}
\alias{test.rkdb}
\title{Run simple series of tests for qserver
Note: Start kdb+ instance on localhost:5000 before running this function}
\usage{
test.rkdb()
}
\description{
Run simple series of tests for qserver
Note: Start kdb+ instance on localhost:5000 before running this function
}
