% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkdb.R
\name{version}
\alias{version}
\title{Version of c api this library is linked to.}
\usage{
version()
}
\value{
integer in a format yyyymmdd.
}
\description{
Version of c api this library is linked to.
}
\examples{
version()
}
