% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkdb.R
\name{execute}
\alias{execute}
\title{Execute \code{query} using \code{con} connection to kdb+.}
\usage{
execute(con, query)
}
\arguments{
\item{con}{Connection handle.}

\item{query}{A string to send to kdb+.}
}
\value{
Result of execution.
}
\description{
Execute \code{query} using \code{con} connection to kdb+.
}
\examples{
\dontrun{
execute(h,"til 10")
execute(h,"dev 1000?0")
}
}
