% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkdb.R
\name{open_connection}
\alias{open_connection}
\title{Connect to kdb+ instance.}
\usage{
open_connection(host = "localhost", port = 5000, user = "",
  timeout = 0, tls = FALSE)
}
\arguments{
\item{host}{Hostname. For Unix Domain sockets pass '0.0.0.0'}

\item{port}{Port number.}

\item{user}{Username and password as user:password string}

\item{timeout}{Number of milliseconds to wait for connection}

\item{tls}{Whether or not to use secure connection.}
}
\value{
Handle to kdb+ instance for \code{execute} and \code{close_connection}.
}
\description{
Connect to kdb+ instance.
}
\examples{
\dontrun{
h<-open_connection()
h<-open_connection(port=5000)
h<-open_connection(port=5000, tls=TRUE)
}
}
